aws-lambda-container/python/requirements.txt
boto3
solana
python-dotenv
sqlalchemy==1.4.48
loguru
python-dotenv
psycopg2-binary
pydantic
base58